Move LocatorInterface accessor/mutator to LocatorAwareTrait and use it in Association.
